do not attach sources anymore
The created bundles does not contain class files as "direct" content
but bundles another JAR file.
It seems the some IDEs do not handle the sources of class files that are
not part of the similar compiled file.
So, as long as that sources are not used by the IDE, we don't need to
publish them.

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>